Move default keybindings including shell commands to external tigrc file
The tigrc file is install to ${sysconfdir} by `make install`. To ignore
the default keybindings and load a different set of keybindings, for
example keybindings that do not allow to run commands that modify the
Git repository, use something like:

	TIGRC_SYSTEM=~/.tigrc.safe tig

This also removes the built-in `git gc` keybinding, which was never
really that useful other than scaring new users.

Fixes #235